requirements.txt
psycopg2-binary
pandas
xlsxwriter
matplotlib
